+++ src/traft/op.rs
@@ -19,8 +19,11 @@ use serde::{Deserialize, Serialize};
 use smol_str::SmolStr;
 use sql::ir::operator::ConflictStrategy;
 use std::collections::BTreeMap;
+use tarantool::datetime::Datetime;
 use tarantool::error::{TarantoolError, TarantoolErrorCode};
+use tarantool::ffi::datetime::MP_DATETIME;
 use tarantool::index::IndexType;
+use tarantool::msgpack::ExtStruct;
 use tarantool::session::UserId;
 use tarantool::space::SpaceEngineType;
 
@@ -465,6 +468,13 @@ impl std::fmt::Display for Op {
                             )?;
                         }
                         map.end()
+                    }
+                    Value::Ext(MP_DATETIME, bytes) => {
+                        let res = Datetime::try_from(ExtStruct::new(MP_DATETIME, bytes));
+                        match res {
+                            Ok(datetime) => serializer.serialize_str(&datetime.to_string()),
+                            Err(e) => serializer.serialize_str(&e),
+                        }
                     }
                     other => other.serialize(serializer),
                 }