+++ CHANGELOG.md
@@ -18,6 +18,7 @@ with the `YY.MINOR.MICRO` scheme.
 - New ALTER SYSTEM parameter `governor_check_replication_error` (default: true)
   enables the checking if replication is broken on any instance, in which case
   the instance will be automatically made Offline.
+- New columns `target_state_reason` & `target_state_change_time` in `_pico_instance` system table
 
 ### Fixes