+++ tarantool/CHANGELOG.md
@@ -44,6 +44,7 @@
 
 - `tlua::Object::read` function which is useful for getting borrowed access to
   multiple values on the lua stack
+- `datetime::Datetime::now_utc` constructor
 
 ### Fixed